ch04: Rename Gordle type to Game
