+++ chapter-04/1_2_read_input/gordle/game.go
@@ -9,14 +9,14 @@ import (
 
 const wordLength = 5
 
-// Gordle holds all the information we need to play a game of gordle.
-type Gordle struct {
+// Game holds all the information we need to play a game of gordle.
+type Game struct {
 	reader *bufio.Reader
 }
 
-// New returns a Gordle variable, which can be used to Play!
-func New(reader io.Reader) *Gordle {
-	g := &Gordle{
+// New returns a Game variable, which can be used to Play!
+func New(reader io.Reader) *Game {
+	g := &Game{
 		reader: bufio.NewReader(reader),
 	}
 	fmt.Println("Welcome to Gordle!")
@@ -25,7 +25,7 @@ func New(reader io.Reader) *Gordle {
 }
 
 // Play runs the game.
-func (g *Gordle) Play() {
+func (g *Game) Play() {
 
 	// ask for a valid word
 	attempt := g.ask()
@@ -34,7 +34,7 @@ func (g *Gordle) Play() {
 }
 
 // ask reads input until a valid suggestion is made (and returned).
-func (g *Gordle) ask() []rune {
+func (g *Game) ask() []rune {
 	fmt.Printf("Enter a %d-character guess:\n", wordLength)
 
 	for {
