+++ chapter-04/1_3_isolate_validate/gordle/game_internal_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 )
 
-func TestGordleAsk(t *testing.T) {
+func TestGameAsk(t *testing.T) {
 	tt := map[string]struct {
 		reader *bufio.Reader
 		want   []rune
@@ -32,7 +32,7 @@ func TestGordleAsk(t *testing.T) {
 
 	for name, tc := range tt {
 		t.Run(name, func(t *testing.T) {
-			g := Gordle{reader: tc.reader}
+			g := Game{reader: tc.reader}
 
 			got := g.ask()
 			if !compareRunes(got, tc.want) {
@@ -56,7 +56,7 @@ func compareRunes(s1, s2 []rune) bool {
 	return true
 }
 
-func TestGordleValidateAttempt(t *testing.T) {
+func TestGameValidateAttempt(t *testing.T) {
 	tt := map[string]struct {
 		word     []rune
 		expected error
@@ -81,7 +81,7 @@ func TestGordleValidateAttempt(t *testing.T) {
 
 	for name, tc := range tt {
 		t.Run(name, func(t *testing.T) {
-			g := &Gordle{}
+			g := &Game{}
 
 			err := g.validateAttempt(tc.word)
 			if !errors.Is(err, tc.expected) {
