+++ chapter-04/1_4_check_for_victory/gordle/game.go
