+++ chapter-04/4_config_options/gordle/configuration.go
@@ -5,12 +5,12 @@ import (
 	"io"
 )
 
-// ConfigFunc defines a configuration function for the Gordle.
-type ConfigFunc func(g *Gordle) error
+// ConfigFunc defines a configuration function for the Game.
+type ConfigFunc func(g *Game) error
 
 // WithReader provider a specific reader from which the player suggestions will be read. Default value is stdin.
 func WithReader(reader io.Reader) ConfigFunc {
-	return func(g *Gordle) error {
+	return func(g *Game) error {
 		g.reader = bufio.NewReader(reader)
 		return nil
 	}
@@ -18,7 +18,7 @@ func WithReader(reader io.Reader) ConfigFunc {
 
 // WithSolution sets the solution to the current game (default is random from corpus).
 func WithSolution(solution []rune) ConfigFunc {
-	return func(g *Gordle) error {
+	return func(g *Game) error {
 		g.solution = solution
 		return nil
 	}
@@ -26,7 +26,7 @@ func WithSolution(solution []rune) ConfigFunc {
 
 // WithMaxAttempts changes the maximum number of attempts (default is unlimited)
 func WithMaxAttempts(maxAttempts int) ConfigFunc {
-	return func(g *Gordle) error {
+	return func(g *Game) error {
 		g.maxAttempts = maxAttempts
 		return nil
 	}
