+++ chapter-04/4_config_options/gordle/game_internal_test.go
@@ -7,14 +7,14 @@ import (
 	"testing"
 )
 
-func TestGordleAsk(t *testing.T) {
+func TestGameAsk(t *testing.T) {
 	tt := map[string]struct {
 		reader *bufio.Reader
 		want   []rune
 	}{
 		"5 characters in english": {
-			reader: bufio.NewReader(strings.NewReader("hello")),
-			want:   []rune("hello"),
+			reader: bufio.NewReader(strings.NewReader("HELLO")),
+			want:   []rune("HELLO"),
 		},
 		"5 characters in arabic": {
 			reader: bufio.NewReader(strings.NewReader("مرحبا")),
@@ -24,7 +24,7 @@ func TestGordleAsk(t *testing.T) {
 			reader: bufio.NewReader(strings.NewReader("こんにちは")),
 			want:   []rune("こんにちは"),
 		},
-		"3 characters in japanese": {
+		"3 and then 5 characters in japanese": {
 			reader: bufio.NewReader(strings.NewReader("こんに\nこんにちは")),
 			want:   []rune("こんにちは"),
 		},
@@ -32,7 +32,10 @@ func TestGordleAsk(t *testing.T) {
 
 	for name, tc := range tt {
 		t.Run(name, func(t *testing.T) {
-			g := Gordle{reader: tc.reader, solution: tc.want}
+			g := Game{
+				reader:          tc.reader,
+				solution:        tc.want,
+				solutionChecker: &solutionChecker{solution: tc.want}}
 
 			got := g.ask()
 			if !compareRunes(got, tc.want) {
@@ -56,7 +59,8 @@ func compareRunes(s1, s2 []rune) bool {
 	return true
 }
 
-func TestGordleValidateAttempt(t *testing.T) {
+func TestGameValidateAttempt(t *testing.T) {
+	g := &Game{solution: []rune("hello")}
 	tt := map[string]struct {
 		word     []rune
 		expected error
@@ -81,8 +85,6 @@ func TestGordleValidateAttempt(t *testing.T) {
 
 	for name, tc := range tt {
 		t.Run(name, func(t *testing.T) {
-			g := &Gordle{solution: []rune("hello")}
-
 			err := g.validateAttempt(tc.word)
 			if !errors.Is(err, tc.expected) {
 				t.Errorf("%c, expected %v, got %v", tc.word, tc.expected, err)
