--- chapter-04/final/Makefile
@@ -1,14 +1,0 @@
-run:
-	go run main.go
-
-build:
-	go build -o bin/wordle main.go
-
-test:
-	go test ./...
-
-cover:
-	go test ./... -cover
-
-benchmark:
-	go test ./... -run=^$ bench=. --benchmem
