--- chapter-04/final/corpus/english.txt
@@ -1,34 +1,0 @@
-TERSE
-CLAIM
-WHOOP
-CREPT
-ANGLE
-RUMBA
-HUSKY
-SNIDE
-ORGAN
-CHASM
-SAUNA
-PURER
-ROBOT
-MAGMA
-GIVEN
-BRAVO
-CLOUD
-ROUSE
-COCOA
-GREED
-PRIMO
-SLANT
-FRESH
-INTER
-ANGST
-TRUNK
-WASTE
-FLUFF
-DEATH
-CATER
-WROTE
-MADAM
-MUSIC
-GAMES
