--- chapter-04/final/go.mod
@@ -1,3 +1,0 @@
-module github.com/ablqk/tiny-go-projects/chapter-04/final
-
-go 1.19
