--- chapter-04/final/gordle/errors.go
@@ -1,9 +1,0 @@
-package gordle
-
-// corpusError defines a sentinel error.
-type corpusError string
-
-// Error is the implementation of the error interface by corpusError.
-func (e corpusError) Error() string {
-	return string(e)
-}
