--- chapter-04/pre-refactoring/Makefile
@@ -1,11 +1,0 @@
-run:
-	go run main.go
-
-build:
-	go build -o bin/wordle main.go
-
-test:
-	go test ./...
-
-cover:
-	go test ./... -cover
