--- chapter-04/pre-refactoring/corpus_5characters.txt
@@ -1,38 +1,0 @@
-slice
-raven
-never
-fever
-frame
-right
-TERSE
-CLAIM
-WHOOP
-CREPT
-ANGLE
-RUMBA
-HUSKY
-SNIDE
-ORGAN
-CHASM
-SAUNA
-PURER
-ROBOT
-MAGMA
-GIVEN
-BRAVO
-CLOUD
-ROUSE
-COCOA
-GREED
-PRIMO
-SLANT
-FRESH
-INTER
-ANGST
-TRUNK
-WASTE
-FLUFF
-DEATH
-CATER
-WROTE
-MADAM
