--- chapter-04/pre-refactoring/go.mod
@@ -1,3 +1,0 @@
-module tiny-go-projects/chapter04
-
-go 1.19
